Reworked amount screens
