+++ app/src/main/java/com/tari/android/wallet/di/ApplicationComponent.kt
@@ -40,7 +40,6 @@ import com.tari.android.wallet.ui.activity.SplashActivity
 import com.tari.android.wallet.ui.activity.debug.DebugActivity
 import com.tari.android.wallet.ui.activity.home.HomeActivity
 import com.tari.android.wallet.ui.activity.onboarding.OnboardingFlowActivity
-import com.tari.android.wallet.ui.fragment.qr.QRScannerActivity
 import com.tari.android.wallet.ui.activity.restore.WalletRestoreActivity
 import com.tari.android.wallet.ui.activity.settings.DeleteWalletActivity
 import com.tari.android.wallet.ui.activity.tx.TxDetailsActivity
@@ -55,6 +54,7 @@ import com.tari.android.wallet.ui.fragment.onboarding.CreateWalletFragment
 import com.tari.android.wallet.ui.fragment.onboarding.IntroductionFragment
 import com.tari.android.wallet.ui.fragment.onboarding.LocalAuthFragment
 import com.tari.android.wallet.ui.fragment.profile.WalletInfoFragment
+import com.tari.android.wallet.ui.fragment.qr.QRScannerActivity
 import com.tari.android.wallet.ui.fragment.restore.chooseRestoreOption.ChooseRestoreOptionFragment
 import com.tari.android.wallet.ui.fragment.restore.chooseRestoreOption.ChooseRestoreOptionViewModel
 import com.tari.android.wallet.ui.fragment.restore.enterRestorationPassword.EnterRestorationPasswordViewModel
@@ -67,6 +67,7 @@ import com.tari.android.wallet.ui.fragment.send.addRecepient.AddRecipientFragmen
 import com.tari.android.wallet.ui.fragment.send.addRecepient.AddRecipientViewModel
 import com.tari.android.wallet.ui.fragment.send.finalize.FinalizeSendTxFragment
 import com.tari.android.wallet.ui.fragment.send.finalize.FinalizeSendTxViewModel
+import com.tari.android.wallet.ui.fragment.send.requestTari.RequestTariViewModel
 import com.tari.android.wallet.ui.fragment.settings.allSettings.AllSettingsFragment
 import com.tari.android.wallet.ui.fragment.settings.allSettings.AllSettingsViewModel
 import com.tari.android.wallet.ui.fragment.settings.backgroundService.BackgroundServiceSettingsViewModel
@@ -172,6 +173,7 @@ internal interface ApplicationComponent {
     fun inject(allSettingsViewModel: AllSettingsViewModel)
     fun inject(addRecipientViewModel: AddRecipientViewModel)
     fun inject(finalizeSentTxListViewModel: FinalizeSendTxViewModel)
+    fun inject(requestTariViewModel: RequestTariViewModel)
     /**
      * Service(s).
      */
@@ -181,5 +183,4 @@ internal interface ApplicationComponent {
     * Broadcast receiver
     */
     fun inject(receiver: BootDeviceReceiver)
-
 }
