+++ app/src/main/java/com/tari/android/wallet/ui/fragment/send/addAmount/AddAmountFragment.kt
@@ -41,7 +41,6 @@ import android.os.*
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
@@ -51,7 +50,7 @@ import com.orhanobut.logger.Logger
 import com.tari.android.wallet.R
 import com.tari.android.wallet.R.color.black
 import com.tari.android.wallet.R.dimen.*
-import com.tari.android.wallet.application.DeepLink
+import com.tari.android.wallet.amountInputBinding.fragment.send.addAmount.keyboard.KeyboardController
 import com.tari.android.wallet.databinding.FragmentAddAmountBinding
 import com.tari.android.wallet.di.DiContainer.appComponent
 import com.tari.android.wallet.infrastructure.Tracker
@@ -68,9 +67,7 @@ import com.tari.android.wallet.util.WalletUtil
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import java.lang.ref.WeakReference
-import java.math.BigInteger
 import javax.inject.Inject
-import kotlin.math.min
 
 /**
  * Amount entry fragment.
@@ -81,44 +78,6 @@ class AddAmountFragment : Fragment(), ServiceConnection {
 
     @Inject
     lateinit var tracker: Tracker
-
-    /**
-     * Maps all the elements (digits & separators) displayed in the amount text to their
-     * corresponding TextViews.
-     */
-    private val elements = mutableListOf<Pair<String, TextView>>()
-
-    /**
-     * Values below are used for scaling up/down of the text size.
-     */
-    private var currentTextSize = 0f
-    private var currentAmountGemSize = 0f
-    private var currentFirstElementMarginStart = 0
-
-    /**
-     * Whether digit entry animation is running.
-     */
-    private var digitAnimIsRunning = false
-
-    /**
-     * Minimum amount is micro Tari.
-     */
-    private val maxNoOfDecimalPlaces = 6
-    private val thousandsGroupSize = 3
-
-    /**
-     * Wait this long before taking action (validation etc.) on the entered amount.
-     */
-    private val actionWaitLengthMs = 500L
-
-    private val decimalSeparator = WalletUtil.amountFormatter.decimalFormatSymbols.decimalSeparator.toString()
-    private val thousandsSeparator = WalletUtil.amountFormatter.decimalFormatSymbols.groupingSeparator.toString()
-
-    /**
-     * Below two are related to amount check and validation.
-     */
-    private val amountCheckHandler = Handler(Looper.getMainLooper())
-    private val amountCheckRunnable = AmountCheckRunnable()
 
     private lateinit var listenerWR: WeakReference<Listener>
 
@@ -136,6 +95,8 @@ class AddAmountFragment : Fragment(), ServiceConnection {
      * Formats the summarized emoji id.
      */
     private lateinit var emojiIdSummaryController: EmojiIdSummaryViewController
+
+    private var keyboardController: KeyboardController = KeyboardController()
 
     private lateinit var walletService: TariWalletService
 
@@ -185,20 +146,14 @@ class AddAmountFragment : Fragment(), ServiceConnection {
     }
 
     private fun setupUI() {
+        keyboardController.setup(requireContext(), AmountCheckRunnable(), ui.numpad, ui.amount)
         recipientUser = arguments?.getParcelable("recipientUser")
-        ui.decimalPointButton.text = decimalSeparator
-        currentTextSize = dimen(add_amount_element_text_size)
-        currentAmountGemSize = dimen(add_amount_gem_size)
-        currentFirstElementMarginStart = dimenPx(add_amount_leftmost_digit_margin_start)
-        // setup amount validation
-        amountCheckRunnable.run()
         // hide tx fee
         ui.txFeeContainerView.invisible()
         // hide/disable continue button
         ui.continueButton.invisible()
         ui.disabledContinueButton.visible()
         // add first digit to the element list
-        elements.add(Pair("0", ui.amountElement0TextView))
         val fullEmojiIdListener = object : FullEmojiIdViewController.Listener {
             override fun animationHide(value: Float) {
                 ui.backButton.alpha = 1 - value
@@ -218,44 +173,33 @@ class AddAmountFragment : Fragment(), ServiceConnection {
         fullEmojiIdViewController.fullEmojiId = recipientUser?.publicKey?.emojiId.orEmpty()
         fullEmojiIdViewController.emojiIdHex = recipientUser?.publicKey?.hexString.orEmpty()
         displayAliasOrEmojiId()
-        val amount = arguments?.getDouble(DeepLink.PARAMETER_AMOUNT, Double.MIN_VALUE)
-        if (isFirstLaunch && amount != null && amount != Double.MIN_VALUE) {
-            val handler = Handler(Looper.getMainLooper())
-            amount.toString().withIndex().forEach { (index, char) ->
-                handler.postDelayed({
-                    if (Character.isDigit(char)) {
-                        onDigitOrSeparatorClicked(char.toString())
-                    } else {
-                        onDigitOrSeparatorClicked(decimalSeparator)
-                    }
-                }, (index + 1) * Constants.UI.AddAmount.numPadDigitEnterAnimDurationMs * 2)
-            }
-            handler.postDelayed(
-                this::setActionBindings,
-                (amount.toString().length + 1) * Constants.UI.AddAmount.numPadDigitEnterAnimDurationMs * 2
-            )
-        } else {
-            setActionBindings()
-        }
+//        val amount = arguments?.getDouble(DeepLink.PARAMETER_AMOUNT, Double.MIN_VALUE)
+//        if (isFirstLaunch && amount != null && amount != Double.MIN_VALUE) {
+//            val handler = Handler(Looper.getMainLooper())
+//            amount.toString().withIndex().forEach { (index, char) ->
+//                handler.postDelayed({
+//                    if (Character.isDigit(char)) {
+//                        onDigitOrSeparatorClicked(char.toString())
+//                    } else {
+//                        onDigitOrSeparatorClicked(decimalSeparator)
+//                    }
+//                }, (index + 1) * Constants.UI.AddAmount.numPadDigitEnterAnimDurationMs * 2)
+//            }
+//            handler.postDelayed(
+//                this::setActionBindings,
+//                (amount.toString().length + 1) * Constants.UI.AddAmount.numPadDigitEnterAnimDurationMs * 2
+//            )
+//        } else {
+//            setActionBindings()
+//        }
+        setActionBindings()
     }
 
     private fun setActionBindings() {
         ui.backButton.setOnClickListener { onBackButtonClicked(it) }
         ui.emojiIdSummaryContainerView.setOnClickListener { emojiIdClicked() }
         ui.txFeeDescTextView.setOnClickListener { onFeeViewClick() }
-        ui.deleteButton.setOnClickListener { deleteButtonClicked() }
         ui.continueButton.setOnClickListener { continueButtonClicked() }
-        ui.pad0Button.setOnClickListener { onDigitOrSeparatorClicked("0") }
-        ui.pad1Button.setOnClickListener { onDigitOrSeparatorClicked("1") }
-        ui.pad2Button.setOnClickListener { onDigitOrSeparatorClicked("2") }
-        ui.pad3Button.setOnClickListener { onDigitOrSeparatorClicked("3") }
-        ui.pad4Button.setOnClickListener { onDigitOrSeparatorClicked("4") }
-        ui.pad5Button.setOnClickListener { onDigitOrSeparatorClicked("5") }
-        ui.pad6Button.setOnClickListener { onDigitOrSeparatorClicked("6") }
-        ui.pad7Button.setOnClickListener { onDigitOrSeparatorClicked("7") }
-        ui.pad8Button.setOnClickListener { onDigitOrSeparatorClicked("8") }
-        ui.pad9Button.setOnClickListener { onDigitOrSeparatorClicked("9") }
-        ui.decimalPointButton.setOnClickListener { onDigitOrSeparatorClicked(decimalSeparator) }
     }
 
     override fun onAttach(context: Context) {
@@ -279,11 +223,6 @@ class AddAmountFragment : Fragment(), ServiceConnection {
         ui.emojiIdSummaryContainerView.visible()
         emojiIdSummaryController.display(emojiId)
         ui.titleTextView.gone()
-    }
-
-    override fun onDetach() {
-        amountCheckHandler.removeCallbacks(amountCheckRunnable)
-        super.onDetach()
     }
 
     private fun onBackButtonClicked(view: View) {
@@ -311,400 +250,6 @@ class AddAmountFragment : Fragment(), ServiceConnection {
         ).show()
     }
 
-    /**
-     * -1 if no decimal separator, index otherwise.
-     */
-    private val decimalSeparatorIndex: Int
-        get() = elements.map { it.first }.indexOf(decimalSeparator)
-
-
-    /**
-     * Inserts thousands separators.
-     *
-     * @return delta width introduced by the removal or addition of thousands separators
-     */
-    private fun updateThousandsSeparators(): Int {
-        var deltaWidth = 0
-        // clear existing thousands separators
-        val itemsToRemove = mutableListOf<Pair<String, TextView>>()
-        for (element in elements.iterator()) {
-            if (element.first == thousandsSeparator) {
-                ui.elementContainerView.removeView(element.second)
-                itemsToRemove.add(element)
-                deltaWidth -= element.second.width
-            }
-        }
-        elements.removeAll(itemsToRemove)
-
-        // insert new thousands separators
-        val decimalIndex = decimalSeparatorIndex
-        val onesEndIndex = if (decimalIndex >= 1) { // no thousands
-            if (decimalIndex < (thousandsGroupSize + 1)) {
-                return deltaWidth
-            } else {
-                decimalIndex
-            }
-        } else {
-            elements.size
-        }
-        for (index in (onesEndIndex - 1) downTo 1) {
-            if ((onesEndIndex - index) % thousandsGroupSize == 0) {
-                val element = insertThousandsSeparator(index)
-                element.second.measure(0, 0)
-                deltaWidth += element.second.measuredWidth
-            }
-        }
-        return deltaWidth
-    }
-
-    /**
-     * Calculates the current amount in microtaris.
-     */
-    private val currentAmount: MicroTari
-        get() {
-            var numberOfZerosToAppend = 6 // for millions
-            val decimalIndex = decimalSeparatorIndex
-            if (decimalIndex > 0) {
-                numberOfZerosToAppend -= (elements.size - decimalIndex - 1)
-            }
-            val amountStringBuilder = StringBuilder()
-            for (element in elements) {
-                if (element.first != thousandsSeparator && element.first != decimalSeparator) {
-                    amountStringBuilder.append(element.first)
-                }
-            }
-            for (i in 0 until numberOfZerosToAppend) {
-                amountStringBuilder.append("0")
-            }
-            return MicroTari(BigInteger(amountStringBuilder.toString()))
-        }
-
-    /**
-     * Digit or separator clicked.
-     */
-    private fun onDigitOrSeparatorClicked(digit: String) {
-        if (digitAnimIsRunning) return
-        // check if entering first digit
-        var enteringFirstDigit = false
-        if (elements.size == 1 && elements[0].first == "0") {
-            // entering first digit as 0 & it's already 0 - return
-            if (digit == "0") return
-            enteringFirstDigit = (digit != decimalSeparator)
-        }
-        // check maximum number of decimal places
-        for ((index, element) in elements.iterator().withIndex()) {
-            if (element.first == decimalSeparator) {
-                val noOfDecimalPlaces = elements.size - index - 1
-                if (noOfDecimalPlaces == maxNoOfDecimalPlaces) {
-                    return
-                }
-                break
-            }
-        }
-        // exit if entering decimal point when a decimal point exists
-        val enteringDecimalPoint = (digit == decimalSeparator)
-        if (enteringDecimalPoint && decimalSeparatorIndex > 0) {
-            return
-        }
-
-        // checks successful - remove any pending validation check
-        amountCheckHandler.removeCallbacks(amountCheckRunnable)
-        // set the running flag
-        digitAnimIsRunning = true
-        // current content width
-        var contentWidthPreInsert =
-            ui.elementContainerView.getLastChild()!!.right - ui.elementContainerView.getFirstChild()!!.left
-        // create new text view
-        val textView = appendAndGetNewDigitTextView(
-            enteringFirstDigit,
-            enteringDecimalPoint,
-            digit
-        )
-        // measure it
-        textView.measure(
-            View.MeasureSpec.UNSPECIFIED,
-            View.MeasureSpec.UNSPECIFIED
-        )
-        val measuredWidth = textView.measuredWidth
-
-        // update thousands separators & get delta width
-        contentWidthPreInsert += updateThousandsSeparators()
-
-        // adjust text sizes if content overflows
-        scaleDownIfRequired(
-            contentWidthPreInsert,
-            enteringFirstDigit,
-            enteringDecimalPoint,
-            textView
-        )
-        // no animation if entering decimal point
-        if (enteringDecimalPoint) {
-            digitAnimIsRunning = false
-            return
-        }
-        // otherwise animate the new digit
-        animateNewDigit(
-            enteringFirstDigit,
-            measuredWidth,
-            textView,
-            digit
-        )
-
-        // setup amount validation
-        amountCheckHandler.postDelayed(
-            amountCheckRunnable,
-            actionWaitLengthMs
-        )
-    }
-
-    /**
-     * Inserts the new element after a num-pad click.
-     * @return the text view for the new digit
-     */
-    private fun appendAndGetNewDigitTextView(
-        enteringFirstDigit: Boolean,
-        enteringDecimalPoint: Boolean,
-        digit: String
-    ): TextView {
-        val inflater = LayoutInflater.from(context)
-        return if (enteringFirstDigit) { // first digit :: use the existing view
-            elements[0].second
-        } else { // inflate text view
-            val textView = inflater.inflate(
-                R.layout.add_amount_element,
-                ui.elementContainerView,
-                false
-            ) as TextView
-            textView.setTextSizePx(currentTextSize)
-            if (!enteringDecimalPoint && !enteringFirstDigit) {
-                textView.alpha = 0f
-                textView.setLayoutWidth(0)
-            } else {
-                textView.setWidthAndHeightToMeasured()
-            }
-            textView.text = digit
-            // add child
-            ui.elementContainerView.addView(textView, ui.elementContainerView.childCount - 1)
-            // add new element
-            elements.add(Pair(digit, textView))
-            textView
-        }
-    }
-
-    /**
-     * Inserts a thousands separator at the specified index.
-     *
-     * @return inserted element
-     */
-    private fun insertThousandsSeparator(index: Int): Pair<String, TextView> {
-        val inflater = LayoutInflater.from(context)
-        val textView = inflater.inflate(
-            R.layout.add_amount_element,
-            ui.elementContainerView,
-            false
-        ) as TextView
-        textView.setTextSizePx(currentTextSize)
-        textView.setWidthAndHeightToMeasured()
-        textView.text = thousandsSeparator
-        // add child - add 1 for the gem icon (it's the very first child)
-        ui.elementContainerView.addView(textView, index + 1)
-        val element = Pair(thousandsSeparator, textView)
-        elements.add(index, element)
-        return element
-    }
-
-    /**
-     * Scales the elements down if the amount overflows after adding the new digit.
-     */
-    private fun scaleDownIfRequired(
-        contentWidthPreInsert: Int,
-        enteringFirstDigit: Boolean,
-        enteringDecimalPoint: Boolean,
-        textView: TextView
-    ) {
-        val contentWidthPostInsert = contentWidthPreInsert + textView.measuredWidth
-        // calculate scale factor
-        var scaleFactor = 1f
-        while ((contentWidthPostInsert * scaleFactor) > ui.elementContainerView.width) {
-            scaleFactor *= 0.95f
-        }
-        currentTextSize *= scaleFactor
-        currentAmountGemSize *= scaleFactor
-        currentFirstElementMarginStart =
-            (currentFirstElementMarginStart * scaleFactor).toInt()
-
-        // adjust gem size
-        ui.amountGemImageView.setLayoutSize(
-            currentAmountGemSize.toInt(),
-            currentAmountGemSize.toInt()
-        )
-        // adjust first element margin
-        elements[0].second.setStartMargin(currentFirstElementMarginStart)
-        // set center correction view width
-        val width = currentAmountGemSize.toInt() + currentFirstElementMarginStart
-        ui.amountCenterCorrectionView.setLayoutWidth(width)
-        // set text sizes
-        for (element in elements) {
-            element.second.setTextSizePx(currentTextSize)
-            if (enteringFirstDigit
-                || enteringDecimalPoint
-                || element !== elements.last()
-            ) {
-                element.second.setWidthAndHeightToMeasured()
-            } else {
-                textView.setLayoutWidth(0)
-            }
-        }
-    }
-
-    /**
-     * Animates the introduction of a new digit.
-     */
-    private fun animateNewDigit(
-        enteringFirstDigit: Boolean,
-        measuredWidth: Int,
-        textView: TextView,
-        digit: String
-    ) {
-        val inflater = LayoutInflater.from(context)
-        var ghostTextView = inflater.inflate(
-            R.layout.add_amount_element,
-            ui.elementOuterContainerView,
-            false
-        ) as TextView?
-        ghostTextView!!.text = digit
-        ghostTextView.setWidthAndHeightToMeasured()
-        ghostTextView.alpha = 0f
-        ui.elementOuterContainerView.addView(ghostTextView)
-
-        val anim = ValueAnimator.ofFloat(0f, 1f)
-        anim.addUpdateListener { valueAnimator: ValueAnimator ->
-            // update text view
-            val value = min(1f, valueAnimator.animatedValue as Float)
-            if (!enteringFirstDigit) {
-                // horizontal scaling if not entering first digit
-                val width = measuredWidth * value
-                textView.setLayoutWidth(width.toInt())
-            } else {
-                // alpha animation if entering first digit
-                textView.alpha = (1 - value) / 2f
-            }
-
-            val location = IntArray(2)
-            textView.getLocationOnScreen(location)
-            // update ghost text view
-            ghostTextView!!.setTextSizePx((currentTextSize / 2f) + ((currentTextSize / 2f) * value))
-            ghostTextView!!.setWidthAndHeightToMeasured()
-            if (enteringFirstDigit) {
-                ghostTextView!!.setStartMargin(
-                    (location[0] + ((1 - value) * measuredWidth * 0.8f)).toInt()
-                )
-            } else {
-                ghostTextView!!.setStartMargin(
-                    (location[0] + ((1 - value) * measuredWidth * 0.3f)).toInt()
-                )
-            }
-            ghostTextView!!.setTopMargin(
-                ui.elementContainerView.top
-                        + dimenPx(add_amount_element_container_translation_y)
-                        + (ui.elementContainerView.height * (1 - value) * 0.8f).toInt()
-            )
-            ghostTextView!!.alpha = value
-
-            if (value == 1f) {
-                if (enteringFirstDigit) {
-                    textView.text = digit
-                    textView.setWidthAndHeightToMeasured()
-                }
-                textView.alpha = 1f
-                ui.elementOuterContainerView.removeView(ghostTextView)
-                ghostTextView = null
-                val lastIndex = elements.size - 1
-                elements[lastIndex] = elements[lastIndex].copy(first = digit)
-                digitAnimIsRunning = false
-            }
-
-        }
-        anim.duration = Constants.UI.AddAmount.numPadDigitEnterAnimDurationMs
-        anim.interpolator = EasingInterpolator(Ease.SINE_OUT)
-        anim.start()
-
-    }
-
-    private fun deleteButtonClicked() {
-        amountCheckHandler.removeCallbacks(amountCheckRunnable)
-        if (elements.size == 1) { // single digit
-            elements[0] = elements[0].copy(first = "0")
-            elements[0].second.text = elements[0].first
-            amountCheckHandler.postDelayed(
-                amountCheckRunnable,
-                actionWaitLengthMs
-            )
-            return
-        }
-        // remove last element
-        val removed = elements.removeAt(elements.size - 1)
-        ui.elementContainerView.removeView(removed.second)
-        // thousands separators
-        updateThousandsSeparators()
-        // scale
-        scaleUpIfRequired()
-        // setup amount check/validation
-        amountCheckHandler.postDelayed(
-            amountCheckRunnable,
-            actionWaitLengthMs
-        )
-    }
-
-    /**
-     * Scales up the elements after a delete if required.
-     */
-    private fun scaleUpIfRequired() {
-        // check content width
-        val first = ui.elementContainerView.getFirstChild()!!
-        val last = ui.elementContainerView.getLastChild()!!
-        var contentWidth = last.right - first.left
-
-        // scale up if needed
-        var scaleFactor = 1f
-        while (contentWidth < ui.elementContainerView.width
-            && (currentTextSize * scaleFactor) < dimen(add_amount_element_text_size)
-        ) {
-            contentWidth = (contentWidth * 1.05f).toInt()
-            if (contentWidth < ui.elementContainerView.width) {
-                scaleFactor *= 1.05f
-            }
-        }
-        // normalize scale factor if needed
-        if ((currentTextSize * scaleFactor) > dimen(add_amount_element_text_size)) {
-            currentTextSize = dimen(add_amount_element_text_size)
-            currentAmountGemSize = dimen(add_amount_gem_size)
-            currentFirstElementMarginStart = dimenPx(add_amount_leftmost_digit_margin_start)
-        } else {
-            currentTextSize *= scaleFactor
-            currentAmountGemSize *= scaleFactor
-            currentFirstElementMarginStart =
-                (currentFirstElementMarginStart * scaleFactor).toInt()
-        }
-
-        // adjust gem size
-        ui.amountGemImageView.setLayoutSize(
-            currentAmountGemSize.toInt(),
-            currentAmountGemSize.toInt()
-        )
-        // adjust first element margin
-        elements[0].second.setStartMargin(
-            currentFirstElementMarginStart
-        )
-        // set center correction view width
-        val width = currentAmountGemSize.toInt() + currentFirstElementMarginStart
-        ui.amountCenterCorrectionView.setLayoutWidth(width)
-        // set text sizes
-        for (element in elements) {
-            element.second.setTextSizePx(currentTextSize)
-            element.second.setWidthAndHeightToMeasured()
-        }
-    }
 
     private fun continueButtonClicked() {
         ui.continueButton.isClickable = false
@@ -715,7 +260,7 @@ class AddAmountFragment : Fragment(), ServiceConnection {
         val error = WalletError()
         val balanceInfo = walletService.getBalanceInfo(error)
         val fee = estimatedFee
-        val amount = currentAmount
+        val amount = keyboardController.currentAmount
         if (error.code == WalletErrorCode.NO_ERROR && fee != null) {
             if (amount > balanceInfo.availableBalance) {
                 lifecycleScope.launch(Dispatchers.Main) {
@@ -764,7 +309,7 @@ class AddAmountFragment : Fragment(), ServiceConnection {
     }
 
     private fun continueToNote() {
-        listenerWR.get()?.continueToAddNote(this, recipientUser!!, currentAmount)
+        listenerWR.get()?.continueToAddNote(this, recipientUser!!, keyboardController.currentAmount)
     }
 
     /**
@@ -776,7 +321,7 @@ class AddAmountFragment : Fragment(), ServiceConnection {
             val error = WalletError()
 
             // update fee
-            val fee = walletService.estimateTxFee(currentAmount, error)
+            val fee = walletService.estimateTxFee(keyboardController.currentAmount, error)
             estimatedFee = fee
 
             if (error.code != WalletErrorCode.NO_ERROR
@@ -790,7 +335,7 @@ class AddAmountFragment : Fragment(), ServiceConnection {
 
             if (error.code == WalletErrorCode.FUNDS_PENDING
                 || error.code == WalletErrorCode.NOT_ENOUGH_FUNDS
-                || (currentAmount + fee) > availableBalance
+                || (keyboardController.currentAmount + fee) > availableBalance
             ) {
                 showErrorState(error)
             } else {
@@ -806,7 +351,7 @@ class AddAmountFragment : Fragment(), ServiceConnection {
             val showsTxFee: Boolean
             txFeeTextView.text = "+${WalletUtil.amountFormatter.format(fee.tariValue)}"
             // show/hide continue button
-            if (currentAmount.value.toInt() == 0) {
+            if (keyboardController.currentAmount.value.toInt() == 0) {
                 hideContinueButton()
                 showsTxFee = false
             } else {
@@ -870,7 +415,7 @@ class AddAmountFragment : Fragment(), ServiceConnection {
 
             showAvailableBalanceError()
 
-            nudgeAmountView()
+            keyboardController.nudgeAmountView()
 
             showOrHideFeeViewAnimated(true)
         }
@@ -894,30 +439,6 @@ class AddAmountFragment : Fragment(), ServiceConnection {
             gemNotEnoughBalance.imageTintList =
                 ColorStateList.valueOf(ContextCompat.getColor(requireContext(), black))
             availableBalanceTextView.setTextColor(ContextCompat.getColor(requireContext(), black))
-        }
-
-        private fun nudgeAmountView() = with(ui) {
-            // don't allow digit entry during this animation
-            digitAnimIsRunning = true
-            ValueAnimator.ofFloat(0f, 1f).apply {
-                addUpdateListener { valueAnimator: ValueAnimator ->
-                    val value = valueAnimator.animatedValue as Float
-                    // nudge the amount
-                    if (value < 0.5f) { // nudge right for the half of the animation
-                        elementContainerView.translationX =
-                            dimenPx(add_amount_available_balance_error_amount_nudge_distance) * value
-                    } else { // nudge back to original position for the second half
-                        elementContainerView.translationX =
-                            dimenPx(add_amount_available_balance_error_amount_nudge_distance) * (1f - value)
-                    }
-                    if (value == 1f) {
-                        digitAnimIsRunning = false
-                    }
-                }
-                duration = Constants.UI.shortDurationMs
-                interpolator = EasingInterpolator(Ease.CIRC_OUT)
-                start()
-            }
         }
 
         private fun showContinueButtonAnimated() = with(ui) {
@@ -958,5 +479,4 @@ class AddAmountFragment : Fragment(), ServiceConnection {
     }
 
     // endregion
-
 }
