+++ app/src/main/java/com/tari/android/wallet/util/WalletUtil.kt
@@ -35,6 +35,7 @@ package com.tari.android.wallet.util
 import com.tari.android.wallet.application.DeepLink
 import com.tari.android.wallet.application.Network
 import com.tari.android.wallet.data.WalletConfig
+import com.tari.android.wallet.model.MicroTari
 import java.io.File
 import java.math.RoundingMode
 import java.text.DecimalFormat
@@ -80,6 +81,10 @@ internal object WalletUtil {
 
     fun getPublicKeyHexDeepLink(publicKeyHex: String, network: Network): String {
         return "${Constants.Wallet.deepLinkURLPrefix}${network.uriComponent}/${DeepLink.Type.PUBLIC_KEY_HEX.uriComponent}/$publicKeyHex"
+    }
+
+    fun generateFullQrCodeDeepLink(publicKeyHex: String, network: Network, amount: MicroTari, alias: String): String {
+        return "${Constants.Wallet.deepLinkURLPrefix}${network.uriComponent}/${DeepLink.Type.PUBLIC_KEY_HEX.uriComponent}/$publicKeyHex?alias=$alias&amount=${amount.tariValue}"
     }
 
     fun getLogFilesFromDirectory(dirPath: String): List<File> {
