+++ app/src/main/res/drawable/ic_share.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="17dp"
+    android:height="20dp"
+    android:viewportWidth="17"
+    android:viewportHeight="20">
+  <path
+      android:pathData="M8.7067,0.1704L11.0531,2.4485C11.2254,2.5748 11.3406,2.7743 11.3717,3.0068C11.4057,3.2554 11.3396,3.5049 11.1924,3.6924C11.0331,3.8928 10.7756,4.0141 10.5011,4.0141C10.3007,4.0141 10.1084,3.9489 9.9591,3.8317L8.9792,2.8261L8.945,12.5738C8.9514,12.7879 8.8673,13.002 8.7073,13.1701C8.5368,13.3503 8.3054,13.4521 8.073,13.4513C7.8234,13.4491 7.5756,13.3205 7.4108,13.1035C7.2905,12.9452 7.2251,12.753 7.2256,12.5627L7.2587,2.8436L6.1887,3.8307C6.0394,3.9479 5.844,4.0131 5.6407,4.0131C5.3612,4.0131 5.1007,3.8918 4.9444,3.6893C4.6499,3.3075 4.717,2.7522 5.0927,2.4485L7.6267,0.1804C7.799,0.0621 7.9713,0.004 8.1487,0H8.1497C8.354,0 8.5354,0.0581 8.7067,0.1704Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M5.3804,6.8457L0.883,6.8457C0.3961,6.8457 0,7.2417 0,7.7285L0,19.1178C0,19.6046 0.3962,20.0006 0.883,20.0006H15.2578C15.7447,20.0006 16.1408,19.6046 16.1408,19.1178L16.1408,7.7285C16.1408,7.2417 15.7447,6.8457 15.2578,6.8457L10.8311,6.8457V8.6103H14.3758L14.3758,18.2361H1.7651L1.765,8.6103L5.3804,8.6103V6.8457Z"
+      android:fillColor="#ffffff"/>
+</vector>
