+++ app/src/main/res/values/strings.xml
@@ -233,6 +233,11 @@
     <string name="send_tari_subtitle">Send</string>
     <string name="request_tari_subtitle">Request</string>
 
+    <!-- request tari -->
+    <string name="request_tari_generate_qr_code">Generate QR Code</string>
+    <string name="request_tari_share">Share</string>
+    <string name="request_tari_share_text">Hi! Please send me %s Tari(s) using this link: %s</string>
+
     <!-- send tari - add recipient -->
     <string name="add_recipient_title">Send To</string>
     <string name="add_recipient_hint">Enter Emoji ID, Contact Name or Yat</string>
